refactor(tournaments): split tournament dal by responsibility
